// src/generated/client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const BULK_OPERATION_TYPE = {
  LOAN_RECIPIENT_UPLOAD: 'LOAN_RECIPIENT_UPLOAD',
  APPLICATION_UPLOAD: 'APPLICATION_UPLOAD'
} as const

export type BULK_OPERATION_TYPE = (typeof BULK_OPERATION_TYPE)[keyof typeof BULK_OPERATION_TYPE]


export const ROLE = {
  SUDO_ADMIN: 'SUDO_ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN',
  ADMIN: 'ADMIN',
  USER: 'USER',
  DATA_ENTRY: 'DATA_ENTRY',
  COMMITTEE_ADMIN: 'COMMITTEE_ADMIN'
} as const

export type ROLE = (typeof ROLE)[keyof typeof ROLE]


export const GENDER = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  OTHER: 'OTHER'
} as const

export type GENDER = (typeof GENDER)[keyof typeof GENDER]


export const APPLICATION_STATUS = {
  INCOMPLETE: 'INCOMPLETE',
  REGISTERED: 'REGISTERED',
  AI_ASSESSED: 'AI_ASSESSED',
  EVALUATED: 'EVALUATED',
  INTERVIEW_SCHEDULED: 'INTERVIEW_SCHEDULED',
  INTERVIEW_COMPLETED: 'INTERVIEW_COMPLETED',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type APPLICATION_STATUS = (typeof APPLICATION_STATUS)[keyof typeof APPLICATION_STATUS]


export const ASSESSMENT_TYPE = {
  AI_SCREENING: 'AI_SCREENING',
  EVALUATION: 'EVALUATION',
  INTERVIEW: 'INTERVIEW'
} as const

export type ASSESSMENT_TYPE = (typeof ASSESSMENT_TYPE)[keyof typeof ASSESSMENT_TYPE]


export const COMMITTEE_TYPE = {
  EVALUATION: 'EVALUATION',
  MONITORING: 'MONITORING'
} as const

export type COMMITTEE_TYPE = (typeof COMMITTEE_TYPE)[keyof typeof COMMITTEE_TYPE]


export const SMS_STATUS = {
  PENDING: 'PENDING',
  SENT: 'SENT',
  FAILED: 'FAILED'
} as const

export type SMS_STATUS = (typeof SMS_STATUS)[keyof typeof SMS_STATUS]


export const OTP_TYPE = {
  REGISTER: 'REGISTER',
  PASSWORD_RESET: 'PASSWORD_RESET',
  LOGIN_VERIFY: 'LOGIN_VERIFY'
} as const

export type OTP_TYPE = (typeof OTP_TYPE)[keyof typeof OTP_TYPE]